chore : 게시글의 수정, 삭제 인증 서비스 적용
- 게시글의 수정, 삭제를 인증 서비스를 적용하여 작성자일 경우 가능하도록 로직 수정

+++ src/main/kotlin/io/github/potahog/blog/service/PostService.kt
@@ -23,6 +23,12 @@ class PostService (private val postRepository: PostRepository) {
 
     fun update(id: Long, request: PostRequest): PostResponse {
         val post = postRepository.findById(id).orElseThrow{ NotFoundException("Cannot update: Post with id $id does not exist") }
+
+        val currentUser = getCurrentUser()
+        if(post.author.id != currentUser.id){
+            throw IllegalStateException("해당 게시글을 수정할 권한이 없습니다.")
+        }
+
         post.title = request.title
         post.content = request.content
         return postRepository.save(post).toResponse()
@@ -30,6 +36,12 @@ class PostService (private val postRepository: PostRepository) {
 
     fun delete(id: Long) {
         val post = postRepository.findById(id).orElseThrow{ NotFoundException("Cannot delete: Post with id $id does not exist") }
+
+        val currentUser = getCurrentUser()
+        if(post.author.id != currentUser.id){
+            throw IllegalStateException("해당 게시글을 삭제할 권한이 없습니다.")
+        }
+
         postRepository.delete(post)
     }
